+++ Source/Supertalk/SupertalkValue.h
@@ -3,6 +3,7 @@
 #pragma once
 
 #include "CoreMinimal.h"
+#include "SupertalkPlayer.h"
 #include "Engine/DataTable.h"
 #include "SupertalkValue.generated.h"
 
@@ -16,8 +17,15 @@ class SUPERTALK_API USupertalkValue : public UObject
 public:
 	const USupertalkValue* GetResolvedValue(const USupertalkPlayer* Player) const;
 	FText ToResolvedDisplayText(const USupertalkPlayer* Player) const;
+	FString ToResolvedInternalString(const USupertalkPlayer* Player) const;
 
+	// Text meant for display to the user.
 	virtual FText ToDisplayText() const PURE_VIRTUAL(USupertalkValue::ToDisplayText,return FText();)
+
+	// Text meant for passing data around - i.e. object paths. This will eventually be removed
+	// once we don't require string handling to call functions.
+	virtual FString ToInternalString() const PURE_VIRTUAL(USupertalkValue::ToInternalText,return FString();)
+	
 	virtual const USupertalkValue* GetMember(FName MemberName) const PURE_VIRTUAL(USupertalkValue::GetMember,return nullptr;)
 
 	virtual bool IsValueEqualTo(const USupertalkValue* Other) const { return this == Other; }
@@ -36,6 +44,7 @@ public:
 	uint8 bValue : 1;
 
 	virtual FText ToDisplayText() const override;
+	virtual FString ToInternalString() const override;
 	virtual const USupertalkValue* GetMember(FName MemberName) const override;
 
 	virtual bool IsValueEqualTo(const USupertalkValue* Other) const override;
@@ -51,6 +60,7 @@ public:
 	FText Text;
 
 	virtual FText ToDisplayText() const override;
+	virtual FString ToInternalString() const override;
 	virtual const USupertalkValue* GetMember(FName MemberName) const override;
 
 	virtual bool IsValueEqualTo(const USupertalkValue* Other) const override;
@@ -66,6 +76,7 @@ public:
 	FName Variable;
 
 	virtual FText ToDisplayText() const override;
+	virtual FString ToInternalString() const override;
 	virtual const USupertalkValue* GetMember(FName MemberName) const override;
 
 	virtual bool IsValueEqualTo(const USupertalkValue* Other) const override { checkNoEntry(); return false; }
@@ -85,6 +96,7 @@ public:
 	TArray<FName> Members;
 
 	virtual FText ToDisplayText() const override;
+	virtual FString ToInternalString() const override;
 	virtual const USupertalkValue* GetMember(FName MemberName) const override;
 
 	virtual bool IsValueEqualTo(const USupertalkValue* Other) const override { checkNoEntry(); return false; }
@@ -121,6 +133,7 @@ public:
 	TObjectPtr<UObject> Object;
 
 	virtual FText ToDisplayText() const override;
+	virtual FString ToInternalString() const override;
 	virtual const USupertalkValue* GetMember(FName MemberName) const override;
 
 	virtual bool IsValueEqualTo(const USupertalkValue* Other) const override;